fix: int to float64 coz im a dumbass

+++ helper/augmented_matrix_test.go
@@ -7,15 +7,15 @@ import (
 
 func TestAugmentedMatrix(t *testing.T) {
 	equations := []Equation{
-		{expression: []int{81, 27, 9, 3, 1}, value: 12},
-		{expression: []int{16, 8, 4, 2, 1}, value: 7},
+		{expression: []float64{81, 27, 9, 3, 1}, value: 12},
+		{expression: []float64{16, 8, 4, 2, 1}, value: 7},
 	}
 
 	got := CreateAugmentedMatrix(equations)
 
 	want := []AugmentedEquation{
-		{Expression: []int{81, 27, 9, 3, 1, 12}},
-		{Expression: []int{16, 8, 4, 2, 1, 7}},
+		{Expression: []float64{81, 27, 9, 3, 1, 12}},
+		{Expression: []float64{16, 8, 4, 2, 1, 7}},
 	}
 
 	if !reflect.DeepEqual(got, want) {
